Move common test helpers into separate file
